+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## NEXT_RELEASE_VERSION_NUMBER
+## 1.0.2
 
 ### Diff against the base branch